api/requirements.txt
fastapi==0.103.1
uvicorn[standard]==0.23.2
sqlalchemy==2.0.21
asyncpg==0.27.0
python-dotenv==1.0.0
greenlet==2.0.2
pydantic==2.5.3
email-validator==1.3.0
psycopg2-binary==2.9.9
 python-multipart